requirements-dev.txt
model-mommy>=1.5.1,<1.6
freezegun>=0.3,<0.4
django-selenium-clean>=0.3.1,<0.4
beautifulsoup4>=4.4,<5
